try find differences

+++ MPI/mpi1516/nbody/nbody-par_other_method_v1.c
@@ -38,20 +38,18 @@ typedef struct {
 
 int globalStartB;
 int globalStartC;
-forceType *forces;
-forceType *new_forces;
-int bodyCt;
-int old = 0;    /* Flips between 0 and 1 */
-bodyType *bodies;
-int *displs;
-int *forces_per_proc;
-int myid;
+forceType *forces;      /*list of forces per body*/
+forceType *new_forces;  /*ausiliar list of forces (MPI reduce)*/
+int bodyCt;             /*number of bodies*/
+int old = 0;            /* Flips between 0 and 1 */
+bodyType *bodies;       /*list of bodies*/
+int *displs;            /*list of the starting indexes of forces assigned per processor*/
+int *forces_per_proc;   /*list of the number of forces assigned per processor*/
+int myid;               /*MPI process ID*/
 int printed = 0;
-int numprocs;
+int numprocs;           /*number of MPI processes involeved in the computation*/
 MPI_Op mpi_sum;
-int totalNumberOfForcesComputed = 0;
-
-
+double totalNumberOfForcesComputed = 0;
 
 /*  Macros to hide memory layout
 */
